requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==39.0.0
cssselect==1.1.0
fabric==2.7.1
idna==3.4
invoke==1.7.3
lxml==4.9.2
mysql-connector-python==8.0.31
numpy==1.19.5
pandas==1.1.5
paramiko==2.12.0
parcel==0.2.2
parsel==1.6.0
pathlib2==2.3.7.post1
protobuf==3.19.6
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.27.1
selenium==3.141.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.13
w3lib==2.0.1
